star-backend/requirements_minimal.txt
Flask==2.3.3
flask-cors==5.0.0
flask-socketio==5.3.2
flask-caching==1.10.1
redis==5.0.8
python-socketio==5.7.2
azure-cosmos==4.7.0
azure-storage-blob==12.19.0
requests==2.28.1
bcrypt==4.0.1
flask-limiter==3.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
flask-migrate==4.0.5
marshmallow==3.20.1
pyswisseph==2.10.3.2
geopy==2.4.1
cachetools==5.3.0
ephem==4.1.2
gunicorn==23.0.0